chore: rename variable
