+++ packages/features/calAIPhone/AIPhoneServiceRegistry.ts
@@ -1,9 +1,9 @@
-import { AIPhoneServiceProviderType } from "./interfaces/ai-phone-service.interface";
+import { AIPhoneServiceProviderType } from "./interfaces/AIPhoneService.interface";
 import type {
   AIPhoneServiceProvider,
   AIPhoneServiceProviderFactory,
   AIPhoneServiceProviderConfig,
-} from "./interfaces/ai-phone-service.interface";
+} from "./interfaces/AIPhoneService.interface";
 
 /**
  * Configuration for the registry
