+++ packages/features/calAIPhone/providers/retellAI/RetellAIPhoneServiceProvider.test.ts
@@ -9,7 +9,7 @@ import type {
   AIPhoneServiceUpdatePhoneNumberParams,
   AIPhoneServiceUpdateAgentParams,
   AIPhoneServiceUpdateModelParams,
-} from "../../interfaces/ai-phone-service.interface";
+} from "../../interfaces/AIPhoneService.interface";
 import type { AgentRepositoryInterface } from "../interfaces/AgentRepositoryInterface";
 import type { PhoneNumberRepositoryInterface } from "../interfaces/PhoneNumberRepositoryInterface";
 import type { TransactionInterface } from "../interfaces/TransactionInterface";
