+++ packages/features/calAIPhone/providers/retellAI/services/AIConfigurationService.ts
@@ -4,7 +4,7 @@ import type {
   AIPhoneServiceUpdateModelParams,
   AIPhoneServiceProviderType,
   AIPhoneServiceModel,
-} from "../../../interfaces/ai-phone-service.interface";
+} from "../../../interfaces/AIPhoneService.interface";
 import { DEFAULT_BEGIN_MESSAGE, DEFAULT_PROMPT_VALUE } from "../../../promptTemplates";
 import { RetellAIServiceMapper } from "../RetellAIServiceMapper";
 import type {
